% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{valh}
\alias{valh-package}
\alias{valh}
\title{Shortest Paths and Travel Time from OpenStreetMap via an Valhalla API}
\description{
An interface between R and the Valhalla API.\cr
Valhalla is a routing
service based on OpenStreetMap data.
See <https://valhalla.github.io/valhalla/> for more
information. This package enables the computation of routes, trips,
isochrones and
travel distances matrices (travel time and kilometric distance).\cr
}
\note{
This package relies on the usage of a running Valhalla service (tested with
version xxx API).\cr

The package ships a sample dataset of 100 random pharmacies in Berlin
(© OpenStreetMap contributors - \url{https://www.openstreetmap.org/copyright/en}).\cr
The sf dataset uses the projection WGS 84 / UTM zone 34N (EPSG:32634).\cr
The csv dataset uses WGS 84 (EPSG:4326).\cr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/riatelab/valh}
  \item Report bugs at \url{https://github.com/riatelab/valh/issues}
}

}
\author{
\strong{Maintainer}: Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID})

Authors:
\itemize{
  \item Matthieu Viry (\href{https://orcid.org/0000-0002-0693-8556}{ORCID})
}

}
