% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_elevation.R
\name{vl_elevation}
\alias{vl_elevation}
\title{Get Elevation Along a Route}
\usage{
vl_elevation(
  loc,
  sampling_dist = NA,
  val.server = "https://valhalla1.openstreetmap.de/"
)
}
\arguments{
\item{loc}{one (or multiples) point(s) to snap to the street network.
\code{loc} can be: \itemize{
  \item a vector of coordinates (longitude and latitude, WGS 84),
  \item a data.frame of longitudes and latitudes (WGS 84),
  \item a matrix of longitudes and latitudes (WGS 84),
  \item an sfc object of type POINT,
  \item an sf object of type POINT.
}}

\item{sampling_dist}{the distance between each point to sample the elevation
(in meters). Default is 'NA' (no sampling).}

\item{val.server}{the URL of the Valhalla server. Default is the demo server
(https://valhalla1.openstreetmap.de/).}
}
\value{
An sf POINT object is returned with the following fields: 'distance'
(the distance from the first points), 'height' (the sampled height on the DEM)
and 'geometry' (the geometry of the sampled point).
}
\description{
Build and send a Valhalla API query to get the elevation at a
set of input locations.
This function interfaces with the \emph{height} service.\cr
If \code{sampling_dist} is provided, the elevation is sampled at regular
intervals along the input locations.
}
