% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_status.R
\name{vl_status}
\alias{vl_status}
\title{Get Valhalla Service Status}
\usage{
vl_status(server = getOption("valh.server"), verbose = FALSE)
}
\arguments{
\item{server}{URL of the Valhalla server.}

\item{verbose}{if TRUE and if the service has it enabled,
it will return additional information about the loaded tileset.}
}
\value{
A list with information on the Valhalla service is returned.
}
\description{
Use this function to return information on the Valhalla server (version etc.).\cr
This function interfaces with the \emph{Status} Valhalla service.
}
\examples{
\dontrun{
vl_status("https://valhalla1.openstreetmap.de/", verbose = FALSE)
}
}
